week2 | Laboratory Activity 2
One of the problems encountered is how to get the title, author, and price in class Book. To access these objects in class Book, their method or function must be in public modifier so they can be accessed by other classes and use it. And last to follow the output in the screenshot i need to put the functions in proper sequence in line 81 to 97.